+++ examples/showoci/showoci_upgrade.sh
@@ -56,7 +56,6 @@ download_file showoci.py
 download_file showoci_data.py
 download_file showoci_service.py
 download_file showoci_output.py
-download_file showoci_csv2adw.py
 download_file CHANGELOG.rst
 download_file README.md
 
@@ -71,5 +70,5 @@ download_file run_daily_report.sh
 ###########################################
 
 echo "" | tee -a $LOG
-echo "Upgrading OCI SDK Drivers - oci oci-cli oracledb" | tee -a $LOG
-python3 -m pip install --upgrade oci oci-cli oracledb pip
+echo "Upgrading OCI SDK Drivers - oci oci-cli" | tee -a $LOG
+python3 -m pip install --upgrade oci oci-cli pip