+++ src/oci/oda/models/update_facebook_channel_details.py
@@ -19,6 +19,10 @@ class UpdateFacebookChannelDetails(UpdateChannelDetails):
         of this class is ``FACEBOOK`` and it should not be changed.
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
+        :param name:
+            The value to assign to the name property of this UpdateFacebookChannelDetails.
+        :type name: str
+
         :param description:
             The value to assign to the description property of this UpdateFacebookChannelDetails.
         :type description: str
@@ -54,6 +58,7 @@ class UpdateFacebookChannelDetails(UpdateChannelDetails):
 
         """
         self.swagger_types = {
+            'name': 'str',
             'description': 'str',
             'type': 'str',
             'session_expiry_duration_in_milliseconds': 'int',
@@ -65,6 +70,7 @@ class UpdateFacebookChannelDetails(UpdateChannelDetails):
         }
 
         self.attribute_map = {
+            'name': 'name',
             'description': 'description',
             'type': 'type',
             'session_expiry_duration_in_milliseconds': 'sessionExpiryDurationInMilliseconds',
@@ -75,6 +81,7 @@ class UpdateFacebookChannelDetails(UpdateChannelDetails):
             'bot_id': 'botId'
         }
 
+        self._name = None
         self._description = None
         self._type = None
         self._session_expiry_duration_in_milliseconds = None